+++ rollup.config.dev.js
@@ -1,0 +1,5 @@
+import {getRollupConfig} from './rollup.config';
+
+const config = getRollupConfig(false);
+
+export default config;